Run unit tests before deploying, use gulp for frontend checks
Tests are (among other reasons) not added as fabfile/test*, because then
the fabfile is opened by the discovery test runner, a thing we do not
want.

Also, this commit removes the functionality to check editable
requirements at the same time. Keep libraries in order yourself (and do
not use editable requirements at all if avoidable).

[#79004058]